test/choreo/requirements.txt
protobuf
bosdyn-choreography-client
